requirements.txt
asgiref==3.5.2
asn1crypto==1.5.1
Django==4.1.2
django-autoslug==1.9.8
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-tinymce==3.5.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
future==0.18.2
idna==3.4
lxml==4.9.1
oscrypto==1.3.0
Pillow==9.2.0
pycparser==2.21
PyJWT==2.5.0
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tinycss2==1.1.1
tzdata==2022.5
uritools==4.0.0
urllib3==1.26.12
webencodings==0.5.1
whitenoise==6.2.0
